Add one button, Refactor code, Restrict all user actions to one activity
