+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         <activity
             android:name=".NavigationActivity"
             android:label="@string/title_activity_navigation" />
-        <activity android:name=".ui.Vector">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
