+++ app/src/main/res/layout/activity_main.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.Vector">
+    tools:context=".ui.MainActivity">
 
     <ImageView
-        android:id="@+id/image_real_life"
+        android:id="@+id/image"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:adjustViewBounds="true"
@@ -18,18 +18,27 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/real_life_back"
+        android:id="@+id/btn_triforce"
         android:layout_width="140dp"
         android:layout_height="wrap_content"
-        android:text="@string/back"
+        android:text="Triforce"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent" />
 
     <Button
-        android:id="@+id/real_life_next"
+        android:id="@+id/btn_vector"
+        android:layout_width="140dp"
+        android:layout_height="wrap_content"
+        android:text="Vector"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toRightOf="@+id/btn_triforce"
+        app:layout_constraintRight_toLeftOf="@+id/btn_real_life" />
+
+    <Button
+        android:id="@+id/btn_real_life"
         android:layout_width="130dp"
         android:layout_height="wrap_content"
-        android:text="@string/next"
+        android:text="Real Life"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintRight_toRightOf="parent" />
